fixed CSV reader and DaS

+++ ImagingDomain.c
@@ -8,12 +8,12 @@ Colton Cox - ccox60@uco.edu
 
 int index = 0;
 
-int generateImagingDomain(double imaging_domain[MAX_SIZE][3])
+int generateImagingDomain(long double imaging_domain[MAX_SIZE][3])
 {
-    for (double x=-0.07;x<0.07;x+=RESOLUTION)
+    for (double x = -0.07; x < 0.07; x += RESOLUTION)
     {
         
-        for (double y=-0.07;y<0.07;y+=RESOLUTION)
+        for (double y = -0.07; y < 0.07; y += RESOLUTION)
         {
             imaging_domain[index][0] = x;
             imaging_domain[index][1] = y;
@@ -38,15 +38,15 @@ int generateImagingDomain(double imaging_domain[MAX_SIZE][3])
     return index;
 }
 
-void writeImageFile(double imaging_domain[MAX_SIZE][3])
+void writeImageFile(long double imaging_domain[MAX_SIZE][3])
 {
     FILE *fp;
     fp = fopen("output.csv", "w+");
     for (int i = 0; i < index; i++)
     {
-        fprintf(fp, "%lf,", imaging_domain[i][0]);
-        fprintf(fp, "%lf,", imaging_domain[i][1]);
-        fprintf(fp, "%lf\n", imaging_domain[i][2]);
+        fprintf(fp, "%Lf,", imaging_domain[i][0]);
+        fprintf(fp, "%Lf,", imaging_domain[i][1]);
+        fprintf(fp, "%Lf\n", imaging_domain[i][2]);
     }
     fclose(fp);
 }
